Awesome \exclamation mark

+++ whoelse.php
@@ -6,6 +6,7 @@
 			اینا میان
 		</title>
 		<link rel="stylesheet" type="text/css" href="css/style.css" />
+		<!-- <link rel="apple-touch-icon" href="/img/apple-touch-icon.png" /> -->
 	</head>
 	<body>
 		<div id="container">
@@ -53,53 +54,66 @@
 						$contents = file_get_contents('db/db.txt');
 						$contents_arr = explode("\n", $contents);
 						$wantedLine_arr = explode(" ", $contents_arr[$number]);
-
-						// echo "contets: " . $contents;
-						// echo "arr: " . $contents_arr . "<br />";
-
-						// foreach ($contents_arr as $con)
-						// {
-						// 	echo $con . "<br />";
-						// }
-						// echo "wanted arr: " . $wantedLine_arr . "<br />";
-						// echo "look at me: " . $wantedLine_arr[0] . "<br />";
-						$wantedLine = $wantedLine_arr[0] . " " . $wantedLine_arr[1] . " " . $wantedLine_arr[2] . " " . $wantedLine_arr[3] . " " . $wantedLine_arr[4] . " " . "yes";
-						$contents_arr[$number] = $wantedLine;
-						file_put_contents('db/db.txt', implode("\n", $contents_arr));
+						if ($wantedLine_arr[5] == "yes") // already said is coming
+						{
+							echo "<span id=\"danger\">";
+							echo "چند بار می‌گی داداچ؟";
+							echo "</span>";
+						}
+						elseif (numberOfComing() >= 15)
+						{
+							echo "<span id=\"danger\">";
+							echo "شرمنده داداچ، پر شده.";
+							echo "</span>";
+						}
+						else
+						{
+							$wantedLine = $wantedLine_arr[0] . " " . $wantedLine_arr[1] . " " . $wantedLine_arr[2] . " " . $wantedLine_arr[3] . " " . $wantedLine_arr[4] . " " . "yes";
+							$contents_arr[$number] = $wantedLine;
+							file_put_contents('db/db.txt', implode("\n", $contents_arr));
+							$numberOfComing++;
+						}
 					}
 					else
 					{
 						$contents = file_get_contents('db/db.txt');
-						$contents_arr = explode("\n", contents);
+						$contents_arr = explode("\n", $contents);
 						$wantedLine_arr = explode(" ", $contents_arr[$number]);
-						$wantedLine = $wantedLine_arr[0] . " " . $wantedLine_arr[1] . " " . $wantedLine_arr[2] . " " . $wantedLine_arr[3] . " " . $wantedLine_arr[4] . " " . "no";
-						$contents_arr[$number] = $wantedLine;
-						file_put_contents('db/db.txt', implode("\n", $contents_arr));
+						if ($wantedLine_arr[5] == "no")
+						{
+							echo "<span id=\"danger\">";
+							echo "اصلاً مگه قرار بود بیای؟ :|";
+							echo "</span>";
+						}
+						else
+						{
+							$wantedLine = $wantedLine_arr[0] . " " . $wantedLine_arr[1] . " " . $wantedLine_arr[2] . " " . $wantedLine_arr[3] . " " . $wantedLine_arr[4] . " " . "no";
+							$contents_arr[$number] = $wantedLine;
+							file_put_contents('db/db.txt', implode("\n", $contents_arr));
+						}
 					}
 				}
 				elseif (!$emptyForm)
 				{
-					echo "<span id=\"wrongPassword\">";
+					echo "<span id=\"danger\">";
 					echo "اشتبا زدی داداچ.";
 					echo "</span>";
 				}
 
-				function changeToYes($line)
+				function numberOfComing()
 				{
-					if (stristr($line, "No"))
+					$sum = 0;
+					$fileContents = file_get_contents('db/db.txt');
+					$contents_arr = explode("\n", $fileContents);
+					// $wantedLine_arr = explode(" ", $contents_arr[$number]);
+					foreach ($contents_arr as $line)
 					{
-						return "replaement line!\n";
+						//echo "line: " . $line;
+						$line_arr = explode(" ", $line);
+						if ($line_arr[5] == "yes")
+							$sum++;
 					}
-					return $data;
-				}
-
-				function changeToNo($data)
-				{
-					if (stristr($data, 'certain word'))
-					{
-						return "replaement line!\n";
-					}
-				return $data;
+					return $sum;
 				}
 			?>
 			<p>
@@ -107,18 +121,21 @@
 			</p>
 			<ol class="people" id="coming">
 				<?php
-					// $isComingFile = fopen("db/isComing.txt", "r");
-					// while (!feof($isComingFile))
-					// {
-					// 	$string = fgets($isComingFile);
-					// 	if ($string != "")
-					// 	{
-					// 		echo "<li>\n" . $string;
-					// 		echo "\n<img src=\"img/yes.png\" class=\"icon\" />";
-					// 		echo "</li>\n";
-					// 	}
-					// }
-					// fclose($isComingFile);
+					$file = fopen("db/db.txt", "r");
+					while (!feof($file))
+					{
+						$line = fgets($file);
+						$line_arr = explode(" ", $line);
+
+						if ($line_arr[5] == "yes\n")
+						{
+							echo "<li>";
+							echo $line_arr[3] . " " . $line_arr[4] . " ";
+							echo "<img src=\"img/yes.png\" class=\"icon\" />";
+							echo  "</li>\n";
+						}
+					}
+					fclose($file);
 				?>
 			</ol>
 			<p>
@@ -126,18 +143,20 @@
 			</p>
 			<ol class="people" id="notComing">
 				<?php
-					// $notComingFile = fopen("db/notComing.txt", "r");
-					// while (!feof($notComingFile))
-					// {
-					// 	$string = fgets($notComingFile);
-					// 	if ($string != "")
-					// 	{
-					// 		echo "<li>\n" . $string;
-					// 		echo "\n<img src=\"img/no.png\" class=\"icon\" />";
-					// 		echo "</li>\n";
-					// 	}
-					// }
-					// fclose($notComingFile);
+					$file = fopen("db/db.txt", "r");
+					while (!feof($file))
+					{
+						$line = fgets($file);
+						$line_arr = explode(" ", $line);
+						if ($line_arr[5] == "no\n")
+						{
+							echo "<li>";
+							echo $line_arr[3] . " " . $line_arr[4] . " ";
+							echo "<img src=\"img/no.png\" class=\"icon\" />";
+							echo  "</li>\n";
+						}
+					}
+					fclose($file);
 				?>
 			</ol>
 			<!--ol class="people" id="notAnswered">
@@ -159,7 +178,7 @@
 			</p>
 		</div>
 		<footer>
-			 &copy; 2015 بهداد (جلوش) و امیر (پشتش)
+			۲۰۱۵ &copy; بهداد
 		</footer>
 	</body>
 </html>
